Refactor notifications to own top-level directory
