+++ ui/package.json
@@ -49,9 +49,11 @@
         "vuex-module-decorators": "^0.11.0"
     },
     "homepage": "https://github.com/jvalue/open-data-service",
-    "licenses": [{
-        "type": "GNU Affero General Public License 3",
-        "url": "https://opensource.org/licenses/AGPL-3.0"
-    }],
+    "licenses": [
+        {
+            "type": "GNU Affero General Public License 3",
+            "url": "https://opensource.org/licenses/AGPL-3.0"
+        }
+    ],
     "repository": "https://github.com/jvalue/open-data-service.git"
 }
