+++ ui/src/datasource/DatasourceEdit.vue
@@ -49,8 +49,8 @@
           </v-stepper-step>
           <v-stepper-content step="2">
             <adapter-config
-              v-bind:isEditMode = "isEditMode"
               v-model="dialogDatasource"
+              :is-edit-mode="isEditMode"
               @validityChanged="validStep2 = $event"
             />
             <stepper-button-group
@@ -149,9 +149,9 @@ const datasourceNamespace = { namespace: 'datasource' }
   components: { AdapterConfig, StepperButtonGroup, DatasourceMetadataConfig, TriggerConfig }
 })
 export default class DatasourceEdit extends Vue {
-  @Action('loadDatasourceById', datasourceNamespace) private loadDatasourceByIdAction!: ( id: number ) => void
-  @Action('createDatasource', datasourceNamespace) private createDatasourceAction!: ( d: Datasource ) => void
-  @Action('updateDatasource', datasourceNamespace) private updateDatsourceAction!: ( d: Datasource ) => void
+  @Action('loadDatasourceById', datasourceNamespace) private loadDatasourceByIdAction!: (id: number) => void
+  @Action('createDatasource', datasourceNamespace) private createDatasourceAction!: (d: Datasource) => void
+  @Action('updateDatasource', datasourceNamespace) private updateDatsourceAction!: (d: Datasource) => void
   @State('selectedDatasource', datasourceNamespace) private selectedDatasource!: Datasource
 
   private isEditMode = false
@@ -187,7 +187,7 @@ export default class DatasourceEdit extends Vue {
       periodic: true,
       firstExecution: new Date(Date.now() + 600000),
       interval: 60000
-    },
+    }
   }
 
   created () {
