+++ ui/src/notification/NotificationEdit.vue
@@ -38,20 +38,20 @@
             <webhook-edit
               v-if="editedNotification.type === 'WEBHOOK'"
               v-model="editedNotification"
-              @validityChanged="validForm = $event"
               style="flex: 1 1 auto"
+              @validityChanged="validForm = $event"
             />
             <firebase-edit
               v-if="editedNotification.type === 'FCM'"
               v-model="editedNotification"
-              @validityChanged="validForm = $event"
               style="flex: 1 1 auto"
+              @validityChanged="validForm = $event"
             />
             <slack-edit
               v-if="editedNotification.type === 'SLACK'"
               v-model="editedNotification"
-              @validityChanged="validForm = $event"
               style="flex: 1 1 auto"
+              @validityChanged="validForm = $event"
             />
           </v-row>
         </v-container>
@@ -81,15 +81,16 @@
 <script lang="ts">
 import Component from 'vue-class-component'
 import Vue from 'vue'
-import NotificationConfig, {WebhookNotification} from '@/pipeline/notifications/notificationConfig'
-import NotificationEditDialog from '@/pipeline/notifications/notificationEditDialog'
 import { Emit } from 'vue-property-decorator'
-import WebhookEdit from '@/pipeline/notifications/WebhookEdit.vue'
-import FirebaseEdit from '@/pipeline/notifications/FirebaseEdit.vue'
-import SlackEdit from '@/pipeline/notifications/SlackEdit.vue'
+
+import NotificationConfig, { WebhookNotification } from '@/notification/notificationConfig'
+import NotificationEditDialog from '@/notification/notificationEditDialog'
+import WebhookEdit from '@/notification/WebhookEdit.vue'
+import FirebaseEdit from '@/notification/FirebaseEdit.vue'
+import SlackEdit from '@/notification/SlackEdit.vue'
 
 @Component({
-  components: {SlackEdit, WebhookEdit, FirebaseEdit }
+  components: { SlackEdit, WebhookEdit, FirebaseEdit }
 })
 export default class PipelineNotifications extends Vue implements NotificationEditDialog {
   private validForm = false;
@@ -105,7 +106,7 @@ export default class PipelineNotifications extends Vue implements NotificationEd
     notificationId: -1,
     condition: 'true',
     url: '',
-    type: "WEBHOOK"
+    type: 'WEBHOOK'
   }
 
   private editedNotification: NotificationConfig = Object.assign({}, this.defaultNotification)
