+++ ui/src/pipeline/router.ts
@@ -1,8 +1,5 @@
-import DatasourceOverview from './DatasourceOverview.vue'
-import DatasourceEdit from './DatasourceEdit.vue'
 import PipelineOverview from './PipelineOverview.vue'
 import PipelineEdit from './PipelineEdit.vue'
-import PipelineNotifications from '@/pipeline/notifications/NotificationOverview.vue'
 
 export default [
   {
@@ -22,11 +19,5 @@ export default [
     name: 'pipeline-edit',
     component: PipelineEdit,
     meta: { title: 'Edit Pipeline', requiresAuth: true, isEditMode: true }
-  },
-  {
-    path: '/pipelines/:pipelineId/notifications',
-    name: 'notification-overview',
-    component: PipelineNotifications,
-    meta: { title: 'View Notificiations', requiresAuth: true }
   }
 ]
